preserved CVS dirs and removing emptied dirs, and enhanced look
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/trunk@2929 b456876b-0849-0410-b77d-98878d47e9d5

+++ tiki/tiki-admin_system.php
@@ -1,5 +1,5 @@
 <?php
-// $Header: /cvsroot/tikiwiki/tiki/tiki-admin_system.php,v 1.2 2003-12-16 14:42:59 mose Exp $
+// $Header: /cvsroot/tikiwiki/tiki/tiki-admin_system.php,v 1.3 2003-12-16 16:27:26 mose Exp $
 
 // Copyright (c) 2002-2003, Luis Argerich, Garland Foster, Eduardo Polidor, et. al.
 // All Rights Reserved. See copyright.txt for details and a complete list of authors.
@@ -7,31 +7,40 @@
 
 require_once ('tiki-setup.php');
 
-function du($location) {
-	if (!$location or !is_dir($location)) return 0;
-	$total = 0;
-	$all = opendir($location);
+function du($path) {
+	if (!$path or !is_dir($path)) return 0;
+	$total = 0; 
+	$cant = 0;
+	$back = array();
+	$all = opendir($path);
 	while ($file = readdir($all)) {
-		if (is_dir($location.'/'.$file) and $file <> ".." and $file <> ".") {
-			$total += du($location.'/'.$file);
+		if (is_dir($path.'/'.$file) and $file <> ".." and $file <> "." and $file <> "CVS") {
+			$du = du($path.'/'.$file);
+			$total+= $du['total'];
+			$cant+= $du['cant'];
 			unset($file);
-		} elseif (!is_dir($location.'/'.$file)) {
-			$stats = stat($location.'/'.$file);
+		} elseif (!is_dir($path.'/'.$file)) {
+			$stats = stat($path.'/'.$file);
 			$total += $stats['size'];
+			$cant++;
 			unset($file);
 		}
 	}
 	closedir($all);
 	unset($all);
-	return $total;
+	$back['total'] = $total;
+	$back['cant'] = $cant;
+	return $back;
 }
 
-function erase_all_dir($path) {
+function erase_dir_content($path) {
+	if (!$path or !is_dir($path)) return 0;
 	if ($dir = opendir($path)) {
 		while (false !== ($file = readdir($dir))) {
-			if (substr($file,0,1) == ".") continue;
+			if (substr($file,0,1) == "." or $file == 'CVS') continue;
 			if (is_dir($path."/".$file)) {
-				erase_all_dir($path."/".$file);
+				erase_dir_content($path."/".$file);
+				rmdir($path."/".$file);
 			} else {
 				unlink($path."/".$file);
 			}
@@ -52,17 +61,17 @@ $buf = '';
 
 if (isset($_GET['do'])) {
 	if ($_GET['do'] == 'templates_c') {
-		erase_all_dir('templates_c');
+		erase_dir_content('templates_c');
 	} elseif ($_GET['do'] == 'modules_cache') {
-		erase_all_dir('modules/cache');
+		erase_dir_content('modules/cache');
 	}
 }
 
-$templates_c_size = du('templates_c');
-$smarty->assign('templates_c_size', $templates_c_size);
+$templates_c = du('templates_c');
+$smarty->assign('templates_c', $templates_c);
 
-$modules_size = du('modules/cache');
-$smarty->assign('modules_size', $modules_size);
+$modules = du('modules/cache');
+$smarty->assign('modules', $modules);
 
 $smarty->assign('mid', 'tiki-admin_system.tpl');
 $smarty->display("tiki.tpl");
